Added using running status in smf_save.c, and its supporter smf_event_can_have_running_status method in smf_decode.c. Also fixed running status handling in smf_load.c.
Remark. An event can have running status if its status byte is the same as
the previous event (except system (realtime and meta, common) messages).

The running status was already handled when loading midi file and now
it is handled also when saving midi file. Thus the result file is
usually shorter than in previous versions.